Much happier with the loading indicator logic
Previously, the loading indicator being displayed was based on whether or not the wallet had had a successful response before. If the first time we open the app, we get a 500 then this flag is never set and we cannot stop the loading indicator from hanging. The logic now is simply a first launch flag. We should the loading indicator on the wallet the first time the user launches and lands on their wallet, regardless of whether they came in via login or not.

5xx status codes are now handled better using the userDriven flag which posts a separate notification to handle the outage silently or to display error.

The loading indicator logic was too complex and wasn’t achieving the basics by trying to do too much.

We can now achieve all the requirements and the code is clean and clear.